finished up CRUD functionality for workouts and users

+++ server/controllers/workoutController.js
@@ -1,10 +1,13 @@
 const asyncHandler = require('express-async-handler')
 const Workout = require('../models/workoutModel')
+const User = require('../models/userModel')
 
 
 // Gets Workouts
 const getWorkout = asyncHandler(async (req, res) => {
-    const workouts = await Workout.find()
+    const workouts = await Workout.find({
+        user: req.user.id
+    })
     res.status(200).json(workouts)
 })
 
@@ -27,7 +30,8 @@ const createWorkout = asyncHandler(async (req, res) => {
     const workout = await Workout.create({
         exercise: req.body.exercise,
         weight: req.body.weight,
-        reps: req.body.reps
+        reps: req.body.reps,
+        user: req.user.id
     })
 
     res.status(200).json({
@@ -44,6 +48,19 @@ const updateWorkout = asyncHandler(async (req, res) => {
         throw new Error('Workout not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // checks to see if logged in user matches the user associated with the workouts
+    if(workout.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedWorkout = await Workout.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     })
@@ -57,6 +74,19 @@ const deleteWorkout = asyncHandler(async (req, res) => {
     if(!workout){
         res.status(400) 
         throw new Error('Workout not found')
+    }
+    
+    const user = await User.findById(req.user.id)
+    
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // checks to see if logged in user matches the user associated with the workouts
+    if(workout.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const deletedWorkout = await Workout.findByIdAndDelete(req.params.id)
